경매 중인 상품에 대한 수정/삭제 예외처리

+++ server/wonprice/src/main/java/main/wonprice/domain/product/controller/ProductController.java
@@ -1,10 +1,12 @@
 package main.wonprice.domain.product.controller;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import main.wonprice.domain.category.entity.Category;
 import main.wonprice.domain.category.service.CategoryService;
 import main.wonprice.domain.member.entity.Member;
 import main.wonprice.domain.member.service.MemberService;
+import main.wonprice.domain.picture.service.PictureService;
 import main.wonprice.domain.product.dto.ProductRequestDto;
 import main.wonprice.domain.product.dto.ProductResponseDto;
 import main.wonprice.domain.product.entity.Product;
@@ -23,12 +25,14 @@ import main.wonprice.domain.product.service.ProductService;
 @Controller
 @RequestMapping("/products")
 @RequiredArgsConstructor
+@Slf4j
 public class ProductController {
 
     private final ProductService productService;
     private final ProductMapper productMapper;
     private final MemberService memberService;
     private final CategoryService categoryService;
+    private final PictureService pictureService;
 
     // 상품 등록
     @PostMapping
@@ -36,6 +40,14 @@ public class ProductController {
         Member loginMember = memberService.findLoginMember();
         Category category = categoryService.findById(productRequestDto.getCategoryId());
         Product product = productService.save(productMapper.toEntity(productRequestDto, loginMember, category));
+
+        /* 대표 */
+//        log.info("productRequestDto.getPictures() : " + productRequestDto.getImages());
+        for (String imageUrl : productRequestDto.getImages()) {
+            pictureService.createPicture(imageUrl, product);
+        }
+        /* 대표 */
+
         ProductResponseDto productResponseDto = productMapper.fromEntity(product);
         return ResponseEntity.ok(productResponseDto);
     }
@@ -101,6 +113,9 @@ public class ProductController {
     @GetMapping("/{productId}")
     public ResponseEntity findOnProduct(@PathVariable Long productId) {
         Product product = productService.findOneById(productId);
+
+        log.info("product : " + product.getProductPictures());
+
         ProductResponseDto productResponseDto = productMapper.fromEntity(product);
         return ResponseEntity.ok(productResponseDto);
     }
@@ -122,15 +137,23 @@ public class ProductController {
     @DeleteMapping("/{productId}")
     public ResponseEntity deleteProduct(@PathVariable Long productId) {
         Member loginMember = memberService.findLoginMember();
-        Product product = productService.deleteOneById(productId, loginMember);
+        Product product = productService.findOneById(productId);
 
         Long productOwnerId = product.getSeller().getMemberId();
         Long loginMemberId = loginMember.getMemberId();
 
         // 상품 판매자와 로그인 한 사용자가 다를 경우, 권한이 없음을 응답
         if (!productOwnerId.equals(loginMemberId)) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("게시글을 삭제할 권한이 없습니다.");
         }
+
+        // 경매 상품인 경우, buyer_id가 있는지 확인하여 삭제 여부 결정
+        if(product.getAuction() && product.getBuyerId() != null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("경매가 진행 중인 상품은 삭제할 수 없습니다.");
+        }
+
+        // 삭제가 가능한 경우, 상품 삭제
+        productService.deleteOneById(productId, loginMember);
 
         return ResponseEntity.noContent().build();
     }
@@ -146,7 +169,12 @@ public class ProductController {
 
         // 상품 판매자와 로그인 한 사용자가 다를 경우, 권한이 없음을 응답
         if (!productOwnerId.equals(loginMemberId)) {
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("게시글을 수정할 권한이 없습니다.");
+        }
+
+        // 경매 중인 상품인 경우, buyer_id가 있는지 확인하여 수정 여부 결정
+        if(updateProduct.getAuction() && updateProduct.getBuyerId() != null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("경매가 진행 중인 상품은 수정할 수 없습니다.");
         }
 
         ProductResponseDto productResponseDto = productMapper.fromEntity(updateProduct);
